NetIf,CRD: Fixed NPE
Apparently the expression var ? then : else forces the else
case to return the same type as the then case. As the then case
already defines a primitive short as a return type, the casted
null will be auto-unboxed into a primitive short, causing an NPE.

Apparently not casting the null to Short prevents the auto-unboxing
and therefore also the NPE.

+++ controller/src/main/java/com/linbit/linstor/core/objects/NetInterfaceDbDriver.java
@@ -92,7 +92,7 @@ public class NetInterfaceDbDriver
                     netIf ->
                     {
                         Integer nullable = TcpPortNumber.getValueNullable(netIf.getStltConnPort(dbCtxRef));
-                        return nullable != null ? nullable.shortValue() : (Short) null;
+                        return nullable != null ? nullable.shortValue() : null;
                     }
                 );
                 break;
